Refactor: 파일삭제 및 에러코드 수정 진행중(#201)

+++ server/controllers/user/logout.js
@@ -1,8 +1,15 @@
-const db = require("../../db");
-const jwt = require("jsonwebtoken");
+const db = require('../../db');
+const jwt = require('jsonwebtoken');
 
 module.exports = async (req, res) => {
   const accessToken = req.cookies.accessToken;
+  if (!accessToken) {
+    return res.status(400).json({
+      data: null,
+      path: '/users/login',
+      message: 'no accessToken',
+    });
+  }
   const verified = jwt.verify(
     accessToken,
     process.env.ACCESS_SECRET,
@@ -11,10 +18,16 @@ module.exports = async (req, res) => {
       return decoded;
     }
   );
-
-  if (verified.isGuest) {
-    await db.query('DELETE FROM users WHERE id=?', [verified.id]);
+  if (verified) {
+    if (verified.isGuest) {
+      await db.query('DELETE FROM users WHERE id=?', [verified.id]);
+    }
+    res.clearCookie('accessToken', { path: '/' });
+    return res.status(204).end();
   }
-  res.clearCookie("accessToken", { path: "/" });
-  res.status(205).json({ message: "로그아웃되었습니다" });
+  return res.status(403).json({
+    data: null,
+    path: '/users/login',
+    message: 'forbidden access',
+  });
 };
